+++ career_ui_legacy/__init__.py
@@ -1,1 +1,0 @@
-"""NiceGUI control plane for Career Workflow."""